langfuse/requirements.txt
llama-index==0.10.18
llama-index-llms-groq==0.1.3
chainlit==1.0.401
groq==0.4.2
python-dotenv==1.0.1
langfuse==2.33.0
llama-index-callbacks-langfuse==0.1.4
